Parameter Make.1-Corpus

type t

type key

val look : t -> length:int -> int -> key option

look data ~length offset extract data of specified length at the given offset. Returns a key that represents this chunk of data (if the data can be extracted).

Note 1 - for simple data representations, like strings, types t and key can be unified, and the look function is just a regular substring extraction function.

Note 2 - the key type is unified with the key of Trie data structure, so one will need to implement the Trie.Key interface.