Module Bap_byteweight.Stats

type t = stats

val trials : t -> int

trial stats is the total number of trials.

This is the total number of occurences of the given substring in all tests, it is equal to h0 stats + h1 stats.

val h0 : t -> int

h0 stats is how many times the null-hypothesis being accepted.

This statistics tells us exactly how many times the label function returned false for this substring.

In terms of the function starts, this is how many times the substring was classified as not a function start.

val h1 : t -> int

h1 stats is how many times the null hypothesis was rejected.

This statistic tells us exactly how many times the label function returned true for this substring.

In terms of the function starts, this is how many times the substring was classified as a function start.