Module Bap_byteweight_signatures

type error = [

| `Corrupted of string	(*
Signature file is corrupted
*)

| `No_signatures	(*
Signature file is not found
*)

| `No_entry of string	(*
Corresponding entry not found
*)

| `Sys_error of string	(*
System error has occurred
*)

]

Error conditions

val save : ?comp:string -> mode:string -> path:string -> Bap.Std.arch -> Regular.Std.bytes -> (unit, error) Core_kernel.Result.t

save ?comp ~mode ~path arch data store signatures data in the database of signatures specified by the path parameter. The triple arch-comp-mode defines a key for the created entry. If an entry with the same name existed, then it would be overwritten with the new data. If the database, doesn't exist, then it will be created and the specified destination.

val load : ?comp:string -> ?path:string -> mode:string -> Bap.Std.arch -> (Regular.Std.bytes, error) Core_kernel.Result.t

load ?comp ?path ~mode arch finds a signature for the specified arch-comp-path triple.

If path is not set, the the signatures are looked up first in default_path and, if not found, in system_path.

Since 2.3.0 the path search has changed to look into two locations.

val default_path : string

default path for the user's signatures database.

Since 2.3.0 it is pointed to the user-specific location, not to the system-wide. See also system_path.

val system_path : string

the path to the system-wide location of signatures.

val string_of_error : error -> string

a human readable representation of an error.