Module Bap_byteweight_signatures.Data

Interface for declaring signature database data types.

val declare : load:( bytes -> 'a ) -> save:( 'a -> bytes ) -> string -> 'a data

declare ~load ~save name declares a new mode.

The load and save functions are used to store the mode information in the signatures database.

Raises an exception if the mode name is not unique.